requirements.txt
appdirs==1.4.4
asgiref==3.4.1
astroid==2.9.1
autopep8==1.6.0
colorama==0.4.4
distlib==0.3.1
Django==4.0
filelock==3.0.12
isort==5.10.1
lazy-object-proxy==1.7.1
mccabe==0.6.1
numpy==1.22.2
pandas==1.4.0
patsy==0.5.2
Pillow==9.0.0
platformdirs==2.4.1
plotly==5.5.0
pycodestyle==2.8.0
pylint==2.12.2
python-dateutil==2.8.2
pytz==2021.3
scipy==1.7.3
six==1.15.0
sqlparse==0.4.2
statsmodels==0.13.1
tenacity==8.0.1
toml==0.10.2
typing_extensions==4.0.1
tzdata==2021.5
virtualenv==20.4.4
virtualenvwrapper-win==1.2.6
wrapt==1.13.3
